добавил api/authApi, services/apiService, constants.tsx
