+++ packages/client/src/constants.tsx
@@ -1,0 +1,1 @@
+export const baseURL = 'https://ya-praktikum.tech/api/v2'
